pyproject.toml
[tool.poetry]
name = "langchainlearning"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
langchain = "^0.0.353"
openai = "^1.7.1"
google-search-results = "^2.4.2"
flask = "^3.0.0"
pinecone-client = "^2.2.4"
tiktoken = "^0.5.2"
pypdf = "^3.17.4"
faiss-cpu = "^1.7.4"
beautifulsoup4 = "^4.12.2"

[tool.poetry.group.dev.dependencies]
black = "^23.12.1"
pandas = "^2.1.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
